+++ romancal/flatfield/flat_field_step.py
@@ -19,16 +19,13 @@ class FlatFieldStep(RomanStep):
     def process(self, step_input):
 
         input_model = rdm.open(step_input)
-        # Get reference file paths
-        reference_file_names = {}
-        try:
-            reffile = self.get_reference_file(input_model, "flat")
-        except CrdsLookupError:
-            reffile = None
-        reference_file_names['flat'] = reffile if reffile != 'N/A' else None
 
-        # Open the relevant reference files as datamodels
-        reference_file_models = {}
+        # Get reference file paths
+        try:
+            reference_file_name = self.get_reference_file(input_model, "flat")
+        except CrdsLookupError:
+            reference_file_name = None
+
 
         # Check for a valid reference file
         if reffile == 'N/A':
@@ -40,20 +37,19 @@ class FlatFieldStep(RomanStep):
             reference_file_models['flat'] = rdm.open(reffile)
             self.log.debug(f'Using FLAT ref file: {reffile}')
         else:
-            reference_file_models['flat'] = None
+            reference_file_model = None
             self.log.debug('Using FLAT ref file')
 
         # Do the flat-field correction
         output_model = flat_field.do_correction(
             input_model,
-            **reference_file_models,
+            reference_file_model,
         )
 
         # Close the input and reference files
         input_model.close()
         try:
-            for model in reference_file_models.values():
-                model.close()
+            reference_file_model.close()
         except AttributeError:
             pass